Add messaging feature
requested by Th!nkPi. Mail command is set-able by rank normally and also
the fake rank "poeticlicense" can be given to any user to authorize them
to use the mail command (for spam prevention measures)
;pl ;upl and ;vpl can all be used to manage the list of users who can
use the command. PL also enables users to use command in PMs (to help
spam the chat less) though the pl commands themselvse have to be used in
chat.
;clearmail can be used by owners and ROs to delete messages for a given
user or to clear the entire inbox for everyone as messages are only
auto-deleted when they are sent.
Inbox size is limited to 5 messages a person, and this can only be
changed in the code itself.
If you have any trouble with all the syntax and whatnot feel free to
contact me about if.

UPDATE: Add timestamps to messages

Writes Date.now() into the messages file so that ";clearmail time"
deletes pending messages that have been existing for a certain amount of
time. (it's currently set to 1 week and this can be adjusted by changing
the MESSAGES_TIME_OUT constant).

Also added ;readmail for if you want to view your inbox without having
the bot see you join the room.