Even more randomness!
Update to randompokemon, now including alternate forms and megas as
possible generation conditions as well as some extra things like shines

Added randomlocation which matches a random adjective from the list
with a random location on the list. My word choice sucks but I'm sure
the writing room can improve on this list!

added randommove which is basically randompokemon but for moves.

added randomcommands as a dummy holder for set so that all the commands
of that nature are set in the same way.