Massive Update!
Changelog:

Non-Code:
I decided to try out Github’s new “Atom” program as opposed to my usual
of Sublime Text. Seems to be going well so far.

Added:
A Choose Your Own Adventure game! Oh, and two commands to go with it
that’ll help with bug fixing, but they’re not important. ^.^’

Changed:
A few grammatical things, and an error that was bugging me.

Removed:
The “;oe” command. The signups have ended, after all. We’re judging it!
^.^