Commands: fix ;time and ;purchase
Alright, time is finally refactored! No more `if` statement hell, and I'd like to think it looks cleaner. It outputs the same thing, though, don't worry.
I also fixed a goofup with ;purchase. In the shop, it says ;buy public greeting, but 'publicgreeting' couldn't actually be purchased because I left it as 'privategreeting'. That should be fixed too.